Add app.java GUI program to repo
Adds the initial draft for the GUI logic that is run out of app.java

Update Event.java to include methods and properties to support the GUI.

Update Seat.java to include properties for row and seat number as
well as floor. Subsequent updates needed in SeatingType and Event to
use the new attributes.

Remove rogue println's in Venue.java.